renamed ChessGame class to WebChessGame
