+++ docs/main.js
@@ -1,5 +1,5 @@
-import { ChessGame } from './chessgame.js';
+import { WebChessGame } from './webchessgame.js';
 const boardContainer = document.querySelector(".board-container");
 if (boardContainer) {
-    new ChessGame(boardContainer);
+    new WebChessGame(boardContainer);
 }
