+++ docs/webchessgame.js
